using workboxto generat eservice-worker.js
The service worker is created with minimal configuration using workbox-cli after the production build is done.
The service worker is always registered; however, it will not register in non-production (npm build) environments because it is only created from the build step.  The registration error is caught and logged to the console.