fixing item edit inputs, combining list and view
making them like the namList inputs: the inputs are only set when the button is clicked
 - deleting Item.newItem()
 - passing props to most handle funcs for simplicity

COMBINING item list and inpum component.  The list is a subcomponent of the Item component.  It is used for View.Item_Read_List.  This allows state to be shared across the read and read_list views